Allow prompt count to have decimals

+++ workflowai/core/domain/llm_usage.py
@@ -6,18 +6,22 @@ from pydantic import BaseModel, Field
 class LLMUsage(BaseModel):
     """An object to store usage information for the LLM provider"""
 
-    prompt_token_count: Optional[int] = Field(
-        default=None, description="The number of tokens in the prompt",
+    prompt_token_count: Optional[float] = Field(
+        default=None,
+        description="The number of tokens in the prompt",
     )
     prompt_cost_usd: Optional[float] = Field(
-        default=None, description="The cost of the prompt in USD",
+        default=None,
+        description="The cost of the prompt in USD",
     )
 
-    completion_token_count: Optional[int] = Field(
-        default=None, description="The number of tokens in the completion",
+    completion_token_count: Optional[float] = Field(
+        default=None,
+        description="The number of tokens in the completion",
     )
     completion_cost_usd: Optional[float] = Field(
-        default=None, description="The cost of the completion in USD",
+        default=None,
+        description="The cost of the completion in USD",
     )
 
     @property
